requirements.txt
aiohttp
pyyaml
playwright
curl_cffi
